+++ windows/attendance_v2/attendance/MainWindow.xaml.cs
@@ -9,6 +9,7 @@ using attendanceManagement.XML;
 using System.IO;
 using MahApps.Metro.Controls.Dialogs;
 using attendanceManagement.widget;
+using attendanceManagement.Models;
 
 
 /***********************************************************************
