+++ windows/attendance_v2/attendance/XML/CurrentCourse.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using attendanceManagement.Models;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
