+++ windows/attendance_v2/attendance/XML/ZXmlDocument.cs
@@ -2,8 +2,7 @@
 using System.IO;
 using System.Collections.Generic;
 using System.Xml;
-
-
+using attendanceManagement.Models;
 
 namespace attendanceManagement.XML
 {
