+++ windows/attendance_v2/attendance/widget/Appbar.xaml.cs
@@ -14,6 +14,7 @@ using System.Windows.Media.Imaging;
 using System.Windows.Navigation;
 using System.Windows.Shapes;
 using MahApps.Metro.Controls;
+using attendanceManagement.Models;
 
 namespace attendanceManagement.widget
 {
