+++ windows/attendance_v2/attendance/widget/CourseSettingDialog.xaml.cs
@@ -12,7 +12,7 @@ using System.Windows.Media;
 using System.Windows.Media.Imaging;
 using System.Windows.Shapes;
 using MahApps.Metro.Controls.Dialogs;
-
+using attendanceManagement.Models;
 
 namespace attendanceManagement.widget
 {
