+++ windows/attendance_v2/attendance/widget/CourseSettingView.xaml.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using attendanceManagement.Models;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Windows;
